requirements.txt
################
# python version 3.6.9
################
scipy==1.2.1
numpy==1.17.0
matplotlib==3.1.2
#opencv_python==4.1.2
#tensorflow==1.15.0
#tensorflow_gpu==1.15.0
#tensorboard==1.15.0
Keras==2.1.5
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0